+++ src/lowRateListDecoder.cpp
@@ -1,7 +1,4 @@
 #include "../include/lowRateListDecoder.h"
-#include "../include/types.h"
-#include "../include/namespace.h"
-#include "../consts.h"
 
 LowRateListDecoder::LowRateListDecoder(FeedForwardTrellis feedforwardTrellis, int listSize, int crcDegree, int crc, char stopping_rule) {
   this->lowrate_nextStates    = feedforwardTrellis.getNextStates();
@@ -695,7 +692,7 @@ std::vector<std::vector<LowRateListDecoder::cell>> LowRateListDecoder::construct
 // converts a path through the tb trellis to the binary message it corresponds with
 std::vector<int> LowRateListDecoder::pathToMessage(std::vector<int> path){
 	std::vector<int> message;
-	for(int pathIndex = 0; pathIndex < path.size() - 1; pathIndex++){
+	for(size_t pathIndex = 0; pathIndex < path.size() - 1; pathIndex++){
 		for(int forwardPath = 0; forwardPath < numForwardPaths; forwardPath++){
 			if(lowrate_nextStates[path[pathIndex]][forwardPath] == path[pathIndex + 1])
 				message.push_back(forwardPath);
@@ -708,7 +705,7 @@ std::vector<int> LowRateListDecoder::pathToMessage(std::vector<int> path){
 // currently does NOT puncture the codeword
 std::vector<int> LowRateListDecoder::pathToCodeword(std::vector<int> path){
 	std::vector<int> nopunc_codeword;
-	for(int pathIndex = 0; pathIndex < path.size() - 1; pathIndex++){
+	for(size_t pathIndex = 0; pathIndex < path.size() - 1; pathIndex++){
 		for(int forwardPath = 0; forwardPath < numForwardPaths; forwardPath++){
 			if(lowrate_nextStates[path[pathIndex]][forwardPath] == path[pathIndex + 1]){
 				std::vector<int> output_bin;
